add test cases
git-svn-id: http://192.168.20.11/sequoiadb/branches/sequoiadb_2.8@32221 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ driver/java/src/test/java/com/sequoiadb/test/SingleCSCLTestCase.java
@@ -25,11 +25,13 @@ public abstract class SingleCSCLTestCase extends SingleCSTestCase {
 
     @AfterClass
     public static void tearDownTestCase() {
-        if (cs.isCollectionExist(clName)) {
-            cs.dropCollection(clName);
+        if (sdb != null && cs != null) {
+            if (cs.isCollectionExist(clName)) {
+                cs.dropCollection(clName);
+            }
+            cl = null;
+            clName = null;
         }
-        cl = null;
-        clName = null;
         SingleCSTestCase.tearDownTestCase();
     }
 }
